refactor(core): rename internal dropdown directives
