+++ projects/core/portals/dropdown/index.ts
@@ -5,10 +5,10 @@ export * from './dropdown.driver';
 export * from './dropdown.providers';
 export * from './dropdown-a11y.directive';
 export * from './dropdown-anchor.directive';
+export * from './dropdown-anchored.directive';
 export * from './dropdown-close.directive';
 export * from './dropdown-content.directive';
 export * from './dropdown-context.directive';
-export * from './dropdown-host.directive';
 export * from './dropdown-hover.directive';
 export * from './dropdown-hover.options';
 export * from './dropdown-limit-width';
